bot_gateway/pyproject.toml
[project]
name = "bot-gateway"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.116.1",
  "uvicorn[standard]==0.35.0",
  "python-telegram-bot==21.6",
  "pydantic==2.11.7",
  "sqlalchemy==2.0.44",
  "httpx==0.28.1",
  "google-genai==1.34.0",
  "redis==5.2.1",
  "rq==1.16.2",
  "opencv-python-headless==4.10.0.84",
  "numpy==2.1.3",
  "matplotlib==3.10.0",
  "pyzbar==0.1.9",
  "pytesseract==0.3.13"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
